Fix encoding input malformed error.
Turns out special characters where missinterpreted and could not be
formatted as I initially wanted them to. The problem was that Chrome has
a default encoding setting which was set to some non-UTF-8 encoding
which messed up some characters for me. So I removed the hacky way of
encoding these characters and updated the README to make sure users add
fix their default encoding instead which should make the rendering more
reliable.